heathbot/lambda/lambda-requirements.txt
discord.py==2.2.2
beautifulsoup4==4.9.0
requests==2.28.2